// src/lib/types.ts
export type Role =
  | 'ROLE_GUEST'
  | 'ROLE_USER'
  | 'ROLE_VERIFIED_USER'
  | 'ROLE_SELLER'
  | 'ROLE_PREMIUM_SELLER'
  | 'ROLE_MODERATOR'
  | 'ROLE_ADMIN'
  | 'ROLE_SUPER_ADMIN';
// ...
export interface UserResponse {
  id: string;
  fullName: string;
  username: string;
  email: string;
  phone?: string | null;
  profileImageUrl?: string | null;
  coverImageUrl?: string | null;
  location?: string | null;
  bio?: string | null;
  roles: Role[];
  emailVerified: boolean;
  phoneVerified: boolean;
  premium: boolean;
  averageRating?: number | null;
  reviewCount?: number | null;
  followersCount?: number | null;
  followingCount?: number | null;
  couponsUploadedCount?: number | null;
  couponsSoldCount?: number | null;
  couponsPurchasedCount?: number | null;
  createdAt: string;
}

export interface AuthResponse {
  accessToken: string;
  refreshToken: string;
  expiresInMs: number;
  user: UserResponse;
}